The authorization manage

+++ app/Http/Controllers/CarController.php
@@ -26,7 +26,12 @@ class CarController extends Controller
      */
     public function create()
     {
-        return view('create');
+        if(auth()->user()->admin) {
+            return view('car.create');
+        }
+        else{
+            abort(403);
+        }
     }
 
     /**
@@ -63,7 +68,7 @@ class CarController extends Controller
      */
     public function show(Car $car)
     {
-        return view('show');
+        
     }
 
     /**
@@ -74,8 +79,12 @@ class CarController extends Controller
      */
     public function edit(Car $car)
     {
-        /*$car = Car::find($car->id);*/
-        return view('car.edit', [ 'car'=>$car ]);
+        if(auth()->user()->admin) {
+            return view('car.edit', [ 'car'=>$car ]);
+        }
+        else{
+            abort(403);
+        }
     }
 
     /**
@@ -101,7 +110,7 @@ class CarController extends Controller
             'plate_number' => $request->plate_number,
             'price_per_day' => $request->price_per_day,
         ]);
-        
+
         return redirect('cars');
     }
 
@@ -113,7 +122,12 @@ class CarController extends Controller
      */
     public function destroy(Car $car)
     {
-        Car::destroy($car->id);
-        return redirect('cars');
+        if(auth()->user()->admin) {
+            Car::destroy($car->id);
+            return redirect('cars');
+        }
+        else{
+            abort(403);
+        }
     }
 }
